fix: fixed the space between sections and the postion of files
Close: #14
