--- src/components/CategoryBox.tsx
@@ -1,5 +1,0 @@
-import { FC } from 'react';
-
-export const dummy: FC = () => (
-  <div />
-);
