+++ src/components/landingPage/OurStoriesSection.tsx
@@ -14,10 +14,10 @@ export const OurStoriesSection: FC = () => (
         xs: '1rem', sm: '5rem', md: '5rem', lg: '5rem', xl: '5rem',
       },
       transform: {
-        xs: 'translateY(135rem)',
-        sm: 'translateY(68rem)',
-        md: 'translateY(60rem)',
-        lg: 'none',
+        xs: 'translateY(138rem)',
+        sm: 'translateY(78rem)',
+        md: 'translateY(70rem)',
+        lg: 'translateY(5rem)',
         xl: 'none',
       },
       mb: '8rem',
