+++ src/components/landingPage/TestimonialsSection.tsx
@@ -21,10 +21,10 @@ export const TestimonialsSection: FC = () => (
       justifyContent: 'center',
       alignItems: 'center',
       transform: {
-        xs: 'translateY(215rem)',
-        sm: 'translateY(175rem)',
-        md: 'translateY(125rem)',
-        lg: 'translateY(30rem)',
+        xs: 'translateY(240rem)',
+        sm: 'translateY(182rem)',
+        md: 'translateY(148rem)',
+        lg: 'translateY(34rem)',
         xl: 'translateY(20rem)',
       },
       mb: '30rem',
